#138-customer navbar
