+++ client/StichHub/src/components/Navbardark.jsx
@@ -1,478 +1,152 @@
-import React, { useState } from "react";
-import { useNavigate } from "react-router-dom";
-import Logo from "../assets/logo/Long - Logo Transparent (White).png";
+import React, { useEffect, useState } from "react";
+import { Link, useLocation, useNavigate } from "react-router-dom";
+import jwt_decode from "jwt-decode";
+import logo from "../assets/logo/Long - Logo Transparent (Black).png";
 
-const pages = ["Home", "Our Services", "How it works", "About", "Contact Us"];
-const settings = ["Profile", "Account", "Dashboard"];
-const setting1 = ["Profile", "Account", "Dashboard", "logout"];
-
-export default function Navbardark() {
-  const [navbar, setNavbar] = useState(false);
-  const [isShown, setIsShown] = useState(false);
-  const [isShown1, setIsShown1] = useState(false);
-  const [isShown2, setIsShown2] = useState(false);
-  const [isShown3, setIsShown3] = useState(false);
-
+const Navbar = () => {
+  const [user, setUser] = useState(JSON.parse(localStorage.getItem("profile")));
+  const [toggle, setToggle] = useState(false);
+  const location = useLocation();
   const navigateTo = useNavigate();
 
-  const handleClick = (event) => {
-    // 👇️ toggle shown state
-    setIsShown((current) => !current);
+  useEffect(() => {
+    const token = user?.token;
 
-    // 👇️ or simply set it to true
-    // setIsShown(true);
-  };
-  const handleClick2 = (event) => {
-    // 👇️ toggle shown state
-    setIsShown1((current) => !current);
+    if (token) {
+      const decodedToken = jwt_decode(token);
 
-    // 👇️ or simply set it to true
-    // setIsShown(true);
-  };
-  const handleClick3 = (event) => {
-    // 👇️ toggle shown state
-    setIsShown2((current) => !current);
+      if (decodedToken.exp * 1000 < new Date().getTime()) logout();
+    }
 
-    // 👇️ or simply set it to true
-    // setIsShown(true);
-  };
-  const handleClick4 = (event) => {
-    // 👇️ toggle shown state
-    setIsShown3((current) => !current);
+    setUser(JSON.parse(localStorage.getItem("profile")));
+  }, [location]);
 
-    // 👇️ or simply set it to true
-    // setIsShown(true);
-  };
-
-  // redirects to /auth/tailor and clears the info of the tailor from localStorage
   const logout = () => {
-    localStorage.clear();
-    navigateTo("/auth/tailor");
+    // localStorage.clear();
+    localStorage.removeItem("profile");
+    navigateTo("/home");
     setUser(null);
   };
 
+  const linkToCart = user ? "/Cart" : "/auth/customer";
+  const linkToMeasurement = user ? "/Cart" : "/auth/customer";
+
   return (
-    <div className="bg-black">
-      <nav
-        className="mx-auto flex max-w-7xl items-center justify-between p-6 lg:px-8"
-        aria-label="Global"
-      >
-        <div className="flex lg:flex-1">
-          <a
-            href="#"
-            className="-m-1.5 p-1.5"
-          >
-            <span className="sr-only">Stichhub</span>
-            <img
-              className="h-8 w-auto"
-              src={Logo}
-              alt=""
-            />
+    <nav className="mt-3">
+      <div className="flex justify-between sm:justify-center">
+        {/* Logo */}
+        <div className="flex mt-2">
+          <a href="/" className="ml-5">
+            <img src={logo} alt="stichHub" className="w-[120px] sm:w-[140px]" />
           </a>
         </div>
-        <a
-          href="#"
-          className="hidden lg:block text-2sm font-semibold leading-6 text-white px-14"
-        >
-          Home
-        </a>
 
-        <div className="flex lg:hidden">
-          <button
-            type="button"
-            className="-m-2.5 inline-flex items-center justify-center rounded-md p-2.5 text-white"
-            onClick={handleClick2}
-          >
-            <span className="sr-only">Open main menu</span>
-            <svg
-              className="h-6 w-6"
-              fill="none"
-              viewBox="0 0 24 24"
-              strokeWidth="1.5"
-              stroke="currentColor"
-              aria-hidden="true"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                d="M3.75 6.75h16.5M3.75 12h16.5m-16.5 5.25h16.5"
-              />
-            </svg>
-          </button>
+        {/* Search */}
+        <div className="hidden sm:flex mx-5 justify-center text-white">
+          <input
+            name="search"
+            id="search"
+            placeholder="search what you want to stitch?"
+            className="border-2 w-[50vw]  rounded-xl border-indigo-500 py-2 px-3 border-solid"
+          />
         </div>
-        <div className="hidden lg:flex lg:gap-x-12">
-          <div className="relative">
-            <button
-              type="button"
-              className="flex items-center gap-x-1 text-2sm font-semibold leading-6 text-white"
-              aria-expanded="false"
-              onClick={handleClick}
-            >
-              Our Services
-              <svg
-                className="h-5 w-5 flex-none text-gray-400"
-                viewBox="0 0 20 20"
-                fill="currentColor"
-                aria-hidden="true"
-              >
-                <path
-                  fillRule="evenodd"
-                  d="M5.23 7.21a.75.75 0 011.06.02L10 11.168l3.71-3.938a.75.75 0 111.08 1.04l-4.25 4.5a.75.75 0 01-1.08 0l-4.25-4.5a.75.75 0 01.02-1.06z"
-                  clipRule="evenodd"
-                />
-              </svg>
-            </button>
 
-            {isShown && (
+        {/* Account Details or signin */}
+        <div className="mx-5">
+          {user?.result ? (
+            <div className="flex ">
+              <span className="mt-2">Welcome! &nbsp;</span>
+              <span className="text-indigo-600 mr-3 mt-2">
+                {" "}
+                {user?.result.name}{" "}
+              </span>
               <div
-                id="navi"
-                className=" absolute -left-8 top-full z-10 mt-3 w-screen max-w-md overflow-hidden rounded-3xl bg-white shadow-lg ring-1 ring-white/5"
+                className="rounded-full w-9 cursor-pointer"
+                onClick={() => setToggle(!toggle)}
               >
-                <div className="p-10">
-                  <div className="group relative flex items-center gap-x-6 rounded-lg p-4 text-sm leading-6 hover:bg-gray-50">
-                    <div className="flex h-11 w-11 flex-none items-center justify-center rounded-lg bg-gray-50 group-hover:bg-white">
-                      <svg
-                        className="h-6 w-6 text-gray-600 group-hover:text-indigo-600"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                        strokeWidth="1.5"
-                        stroke="currentColor"
-                        aria-hidden="true"
-                      >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          d="M10.5 6a7.5 7.5 0 107.5 7.5h-7.5V6z"
-                        />
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          d="M13.5 10.5H21A7.5 7.5 0 0013.5 3v7.5z"
-                        />
-                      </svg>
-                    </div>
-                    <div className="flex-auto">
-                      <a
-                        href="#"
-                        className="block font-semibold text-white"
-                      >
-                        <span className="absolute inset-0"></span>
-                      </a>
-                      <p className="mt-1 text-gray-600"></p>
-                    </div>
+                {user?.result.picture ? (
+                  <img src={user?.result.picture} className="rounded-full" />
+                ) : (
+                  <div className="w-9 bg-indigo-600 rounded-full h-9 text-white text-center text-2xl">
+                    {user?.result.name.charAt(0)}
                   </div>
-
-                  <div className="group relative flex items-center gap-x-6 rounded-lg p-4 text-sm leading-6 hover:bg-gray-50">
-                    <div className="flex h-11 w-11 flex-none items-center justify-center rounded-lg bg-gray-50 group-hover:bg-white">
-                      <svg
-                        className="h-6 w-6 text-gray-600 group-hover:text-indigo-600"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                        strokeWidth="1.5"
-                        stroke="currentColor"
-                        aria-hidden="true"
-                      >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          d="M15.042 21.672L13.684 16.6m0 0l-2.51 2.225.569-9.47 5.227 7.917-3.286-.672zM12 2.25V4.5m5.834.166l-1.591 1.591M20.25 10.5H18M7.757 14.743l-1.59 1.59M6 10.5H3.75m4.007-4.243l-1.59-1.59"
-                        />
-                      </svg>
-                    </div>
-                    <div className="flex-auto">
-                      <a
-                        href="#"
-                        className="block font-semibold text-white"
-                      >
-                        <span className="absolute inset-0"></span>
-                      </a>
-                      <p className="mt-1 text-gray-600"></p>
-                    </div>
-                  </div>
-
-                  <div className="group relative flex items-center gap-x-6 rounded-lg p-4 text-sm leading-6 hover:bg-gray-50">
-                    <div className="flex h-11 w-11 flex-none items-center justify-center rounded-lg bg-gray-50 group-hover:bg-white">
-                      <svg
-                        className="h-6 w-6 text-gray-600 group-hover:text-indigo-600"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                        strokeWidth="1.5"
-                        stroke="currentColor"
-                        aria-hidden="true"
-                      >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          d="M7.864 4.243A7.5 7.5 0 0119.5 10.5c0 2.92-.556 5.709-1.568 8.268M5.742 6.364A7.465 7.465 0 004.5 10.5a7.464 7.464 0 01-1.15 3.993m1.989 3.559A11.209 11.209 0 008.25 10.5a3.75 3.75 0 117.5 0c0 .527-.021 1.049-.064 1.565M12 10.5a14.94 14.94 0 01-3.6 9.75m6.633-4.596a18.666 18.666 0 01-2.485 5.33"
-                        />
-                      </svg>
-                    </div>
-                    <div className="flex-auto">
-                      <a
-                        href="#"
-                        className="block font-semibold text-white"
-                      >
-                        <span className="absolute inset-0"></span>
-                      </a>
-                      <p className="mt-1 text-gray-600"></p>
-                    </div>
-                  </div>
-
-                  <div className="group relative flex items-center gap-x-6 rounded-lg p-4 text-sm leading-6 hover:bg-gray-50">
-                    <div className="flex h-11 w-11 flex-none items-center justify-center rounded-lg bg-gray-50 group-hover:bg-white">
-                      <svg
-                        className="h-6 w-6 text-gray-600 group-hover:text-indigo-600"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                        strokeWidth="1.5"
-                        stroke="currentColor"
-                        aria-hidden="true"
-                      >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          d="M13.5 16.875h3.375m0 0h3.375m-3.375 0V13.5m0 3.375v3.375M6 10.5h2.25a2.25 2.25 0 002.25-2.25V6a2.25 2.25 0 00-2.25-2.25H6A2.25 2.25 0 003.75 6v2.25A2.25 2.25 0 006 10.5zm0 9.75h2.25A2.25 2.25 0 0010.5 18v-2.25a2.25 2.25 0 00-2.25-2.25H6a2.25 2.25 0 00-2.25 2.25V18A2.25 2.25 0 006 20.25zm9.75-9.75H18a2.25 2.25 0 002.25-2.25V6A2.25 2.25 0 0018 3.75h-2.25A2.25 2.25 0 0013.5 6v2.25a2.25 2.25 0 002.25 2.25z"
-                        />
-                      </svg>
-                    </div>
-                    <div className="flex-auto">
-                      <a
-                        href="#"
-                        className="block font-semibold text-white"
-                      >
-                        <span className="absolute inset-0"></span>
-                      </a>
-                      <p className="mt-1 text-gray-600"></p>
-                    </div>
-                  </div>
-
-                  <div className="group relative flex items-center gap-x-6 rounded-lg p-4 text-sm leading-6 hover:bg-gray-50">
-                    <div className="flex h-11 w-11 flex-none items-center justify-center rounded-lg bg-gray-50 group-hover:bg-white">
-                      <svg
-                        className="h-6 w-6 text-gray-600 group-hover:text-indigo-600"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                        strokeWidth="1.5"
-                        stroke="currentColor"
-                        aria-hidden="true"
-                      >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          d="M16.023 9.348h4.992v-.001M2.985 19.644v-4.992m0 0h4.992m-4.993 0l3.181 3.183a8.25 8.25 0 0013.803-3.7M4.031 9.865a8.25 8.25 0 0113.803-3.7l3.181 3.182m0-4.991v4.99"
-                        />
-                      </svg>
-                    </div>
-                    <div className="flex-auto">
-                      <a
-                        href="#"
-                        className="block font-semibold text-white"
-                      >
-                        <span className="absolute inset-0"></span>
-                      </a>
-                      <p className="mt-1 text-gray-600"></p>
-                    </div>
-                  </div>
-                </div>
-                <div className="grid grid-cols-2 divide-x divide-white/5 bg-gray-50">
-                  <a
-                    href="#"
-                    className="flex items-center justify-center gap-x-2.5 p-3 text-sm font-semibold leading-6 text-white hover:bg-gray-100"
-                  >
-                    <svg
-                      className="h-5 w-5 flex-none text-gray-400"
-                      viewBox="0 0 20 20"
-                      fill="currentColor"
-                      aria-hidden="true"
-                    >
-                      <path
-                        fillRule="evenodd"
-                        d="M2 10a8 8 0 1116 0 8 8 0 01-16 0zm6.39-2.908a.75.75 0 01.766.027l3.5 2.25a.75.75 0 010 1.262l-3.5 2.25A.75.75 0 018 12.25v-4.5a.75.75 0 01.39-.658z"
-                        clipRule="evenodd"
-                      />
-                    </svg>
-                    Watch Demo
-                  </a>
-
-                  <a
-                    href="#"
-                    className="flex items-center justify-center gap-x-2.5 p-3 text-sm font-semibold leading-6 text-white hover:bg-gray-100"
-                  >
-                    <svg
-                      className="h-5 w-5 flex-none text-gray-400"
-                      viewBox="0 0 20 20"
-                      fill="currentColor"
-                      aria-hidden="true"
-                    >
-                      <path
-                        fillRule="evenodd"
-                        d="M2 3.5A1.5 1.5 0 013.5 2h1.148a1.5 1.5 0 011.465 1.175l.716 3.223a1.5 1.5 0 01-1.052 1.767l-.933.267c-.41.117-.643.555-.48.95a11.542 11.542 0 006.254 6.254c.395.163.833-.07.95-.48l.267-.933a1.5 1.5 0 011.767-1.052l3.223.716A1.5 1.5 0 0118 15.352V16.5a1.5 1.5 0 01-1.5 1.5H15c-1.149 0-2.263-.15-3.326-.43A13.022 13.022 0 012.43 8.326 13.019 13.019 0 012 5V3.5z"
-                        clipRule="evenodd"
-                      />
-                    </svg>
-                    Support
-                  </a>
-                </div>
+                )}
               </div>
-            )}
-          </div>
 
-          <a
-            href="#"
-            className="text-2sm font-semibold leading-6 text-white"
-          >
-            How it Works?
-          </a>
-          <a
-            href="#"
-            className="text-2sm font-semibold leading-6 text-white"
-          >
-            About
-          </a>
-          <a
-            href="#"
-            className="text-2sm font-semibold leading-6 text-white"
-          >
-            Contact Us
-          </a>
-        </div>
-        <div className="hidden lg:flex lg:flex-1 lg:justify-end">
-        <div
-          onClick={logout}
-          className="cursor-pointer text-2sm  font-semibold leading-6 text-white"
-        >
-          Log Out <span aria-hidden="true">&rarr;</span>
-        </div>
-        </div>
-      </nav>
-      {isShown1 && (
-        <div
-          className="  lg:hidden"
-          role="dialog"
-          aria-modal="true"
-        >
-          <div className="fixed inset-0 z-10"></div>
-          <div className="fixed inset-y-0 right-0 z-10 w-full overflow-y-auto bg-white px-6 py-6 sm:max-w-sm sm:ring-1 sm:ring-white/10">
-            <div className="flex items-center justify-between">
-              <a
-                href="#"
-                className="-m-1.5 p-1.5"
+              {/* User Menu */}
+              <div
+                className={`${
+                  !toggle ? "hidden" : "flex"
+                } cardGradient w-[240px] absolute mx-[-80px] lg:mx-[-10px] py-2 min-w-[140px] top-14 rounded-xl z-[100]`}
               >
-                <span className="sr-only">Stichhub</span>
-                <img
-                  className="h-8 w-auto"
-                  src={Logo}
-                  alt=""
-                />
-              </a>
-              <button
-                type="button"
-                className="-m-2.5 rounded-md p-2.5 text-white-700"
-                onClick={handleClick2}
-              >
-                <span className="sr-only">Close menu</span>
-                <svg
-                  className="h-6 w-6"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  strokeWidth="1.5"
-                  stroke="currentColor"
-                  aria-hidden="true"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="M6 18L18 6M6 6l12 12"
-                  />
-                </svg>
-              </button>
-            </div>
+                <div className="flex flex-col gap-2 mt-3">
+                  <Link to="/Cart">
+                    <span className="ml-6 text-white cursor-pointer my-2 hover:text-cyan-300">
+                      Your Cart
+                    </span>
+                  </Link>
 
-            <div className="mt-6 flow-root">
-              <a
-                href="#"
-                className="-mx-3 block rounded-lg py-2 px-3 text-base font-semibold leading-7 text-white hover:bg-gray-50"
-              >
-                Home
-              </a>
+                  <Link to="/Orders">
+                    <span className="ml-6 text-white cursor-pointer my-2 hover:text-cyan-300">
+                      Your Orders
+                    </span>
+                  </Link>
 
-              <div className="-my-6 divide-y divide-gray-500/10">
-                <div className="space-y-2 py-6">
-                  <div className="-mx-3">
-                    <button
-                      type="button"
-                      onClick={handleClick4}
-                      className="text-white flex w-full items-center justify-between rounded-lg py-2 pl-3 pr-3.5 text-base font-semibold leading-7 hover:bg-gray-50"
-                      aria-controls="disclosure-1"
-                      aria-expanded="false"
-                    >
-                      Our Services
-                      <svg
-                        className="h-5 w-5 flex-none"
-                        viewBox="0 0 20 20"
-                        fill="currentColor"
-                        aria-hidden="true"
-                      >
-                        <path
-                          fillRule="evenodd"
-                          d="M5.23 7.21a.75.75 0 011.06.02L10 11.168l3.71-3.938a.75.75 0 111.08 1.04l-4.25 4.5a.75.75 0 01-1.08 0l-4.25-4.5a.75.75 0 01.02-1.06z"
-                          clipRule="evenodd"
-                        />
-                      </svg>
-                    </button>
-                    {isShown3 && (
-                      <div
-                        className="mt-2 space-y-2"
-                        id="disclosure-1"
-                      >
-                        <a
-                          href="#"
-                          className="block rounded-lg py-2 pl-6 pr-3 text-sm font-semibold leading-7 text-white hover:bg-gray-50"
-                        >
-                          3d Measurement
-                        </a>
+                  <Link to="/home">
+                    <span className="ml-6 text-white cursor-pointer my-2 hover:text-cyan-300">
+                      Your Profile
+                    </span>
+                  </Link>
 
-                        {/* <a href="#" className="block rounded-lg py-2 pl-6 pr-3 text-sm font-semibold leading-7 text-white hover:bg-gray-50">Engagement</a>
-                <a href="#" className="block rounded-lg py-2 pl-6 pr-3 text-sm font-semibold leading-7 text-white hover:bg-gray-50">Security</a>
-                <a href="#" className="block rounded-lg py-2 pl-6 pr-3 text-sm font-semibold leading-7 text-white hover:bg-gray-50">Integrations</a>
-                <a href="#" className="block rounded-lg py-2 pl-6 pr-3 text-sm font-semibold leading-7 text-white hover:bg-gray-50">Automations</a>
-                <a href="#" className="block rounded-lg py-2 pl-6 pr-3 text-sm font-semibold leading-7 text-white hover:bg-gray-50">Watch demo</a>
-                <a href="#" className="block rounded-lg py-2 pl-6 pr-3 text-sm font-semibold leading-7 text-white hover:bg-gray-50">Contact sales</a> */}
-                      </div>
-                    )}
-                  </div>
-                  <a
-                    href="#"
-                    className="-mx-3 block rounded-lg py-2 px-3 text-base font-semibold leading-7 text-white hover:bg-gray-50"
-                  >
-                    How It Works?
-                  </a>
-                  <a
-                    href="#"
-                    className="-mx-3 block rounded-lg py-2 px-3 text-base font-semibold leading-7 text-white hover:bg-gray-50"
-                  >
-                    About
-                  </a>
-                  <a
-                    href="#"
-                    className="-mx-3 block rounded-lg py-2 px-3 text-base font-semibold leading-7 text-white hover:bg-gray-50"
-                  >
-                    Contact US
-                  </a>
-                </div>
-                <div className="py-6">
-                  <div
+                  <Link to="/home">
+                    <span className="ml-6 text-white cursor-pointer my-2 hover:text-cyan-300">
+                      Your Measurements
+                    </span>
+                  </Link>
+                  <button
+                    className="bg-red-500 text-white py-1 px-3 rounded-lg ml-6 mt-9 mb-3"
                     onClick={logout}
-                    className="cursor-pointer text-2sm  font-semibold leading-6 text-white"
                   >
-                      Log Out
-                    </div>  
+                    Log Out
+                  </button>
                 </div>
               </div>
             </div>
-          </div>
+          ) : (
+            <Link to="/auth/customer">
+              <button className="bg-blue-500 text-white py-1 px-3 rounded-lg mt-1">
+                Sign In
+              </button>
+            </Link>
+          )}
         </div>
-      )}
-    </div>
+      </div>
+
+      {/* Search Bar */}
+      <div className="sm:hidden mx-5 flex justify-center mt-1">
+        <input
+          name="search"
+          id="search"
+          placeholder="search what you want to stitch?"
+          className="border-2 w-[90vw] rounded-xl border-indigo-500 py-2 px-3 border-solid"
+        />
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          fill="none"
+          viewBox="0 0 24 24"
+          strokeWidth={1.5}
+          stroke="currentColor"
+          className="w-6 h-6 absolute right-8 mt-2"
+        >
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            d="M21 21l-5.197-5.197m0 0A7.5 7.5 0 105.196 5.196a7.5 7.5 0 0010.607 10.607z"
+          />
+        </svg>
+      </div>
+    </nav>
   );
-}
+};
+
+export default Navbar;
